Change API from imperative to declarative
Instead of exposing a global cometd object that clients interact with
imperativly, we now expose a new type Cometd, that clients instantiate
in their item-hirarchy. This maps to how other non-visual items are
represented in QML, for example Timer and WorkerScript, and allows the
user to easily add multiple instances. It also allows us to add
declarative properties for things such as the configuration.